Updated OCaml

+++ plugins/Full_CI_ZMQ/pt2_stoch_routines.irp.f
@@ -17,7 +17,7 @@ END_PROVIDER
   e = elec_num - n_core_orb * 2
   pt2_n_tasks_max = 1+min((e*(e-1))/2, int(dsqrt(dble(N_det_selectors)))/10)
   do i=1,N_det_generators
-    pt2_F(i) = 1 + dble(pt2_n_tasks_max)*maxval(dsqrt(dabs(psi_coef_sorted_gen(i,1:N_states))))
+    pt2_F(i) = 1 + int(dble(pt2_n_tasks_max)*maxval(dsqrt(dabs(psi_coef_sorted_gen(i,1:N_states)))))
   enddo
   
   if(N_det_generators < 1024) then
